chore(scrapers): extract EventTeam to event_team.py + capture gotsport fixtures
Prep for Shell 01 (gotsport tier-section parser). Extracts the EventTeam
dataclass into its own module so the new pure tier-parser helpers can
import it without dragging the requests/urllib3 stack into the import
graph. Also captures 8 real-event landing pages and 3 synthetic
landings under tests/fixtures/gotsport/ for fixture-driven testing.
