+++ tests/fixtures/gotsport/event_synthetic_single_tier.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html>
+<head><title>Synthetic single-tier event (no ?group= links)</title></head>
+<body>
+<h1>Synthetic Single-Tier Event</h1>
+<script>
+var jsonTeamRegs = [
+  {"id": 1001, "full_name": "Single Tier Team One", "display_age_group": "U13", "display_gender": "Male"},
+  {"id": 1002, "full_name": "Single Tier Team Two", "display_age_group": "U13", "display_gender": "Male"}
+];
+</script>
+<!-- No ?group= links anywhere — source #1 must yield empty. -->
+</body>
+</html>
